+++ react-components/src/pages/types.ts
@@ -1,0 +1,17 @@
+export interface ICardData {
+  name?: string
+  birth?: string
+  category?: string
+  image?: string
+  checkbox?: string
+  switcher?: string
+}
+
+export interface IState {
+  cardsData: ICardData[]
+  errors: IErors
+}
+
+interface IErors {
+  name?: string
+}
